[fix|backend|session][公共]get_session_count 返回 u32 修 specta BigIntForbidden(usize)

+++ src-tauri/src/commands/session.rs
@@ -11,9 +11,9 @@ pub async fn get_sessions(state: State<'_, AppState>) -> Result<Vec<SessionSumma
 
 #[tauri::command]
 #[specta::specta]
-pub async fn get_session_count(state: State<'_, AppState>) -> Result<usize, String> {
+pub async fn get_session_count(state: State<'_, AppState>) -> Result<u32, String> {
     let store = state.store.lock().await;
-    Ok(store.active_count())
+    Ok(store.active_count() as u32)
 }
 
 #[tauri::command]
